+++ protoc_plugin/CHANGELOG.md
@@ -8,6 +8,8 @@
 
   If a target is built with `dart_env = {"protobuf.omit_enum_names": "true"}`
   enum names will not be present in the compiled binary.
+* Make message and field names dependenc on a fromEnvironment constants
+  `protobuf.omit_message_names` and `protobuf.omit_field_names` respectively.
 * Omit type on a left hand side of generated static fields for extensions,
   which results in stricter type (`Extension<ExtensionType>` instead of just
   `Extension`).